+++ src/config.c
@@ -1,5 +1,5 @@
 /*
- *  Copyright (C) 2019-2020 Scoopta
+ *  Copyright (C) 2019-2024 Scoopta
  *  This file is part of Wofi
  *  Wofi is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -81,12 +81,12 @@ char* config_get(struct map* config, const char* key, char* def_opt) {
 	return opt;
 }
 
-uint8_t config_get_mnemonic(struct map* config, const char* key, char* def_opt, uint8_t num_choices, ...) {
+int config_get_mnemonic(struct map* config, const char* key, char* def_opt, int num_choices, ...) {
 	char* opt = config_get(config, key, def_opt);
 	va_list ap;
 	va_start(ap, num_choices);
-	uint8_t result = 0;
-	for(uint8_t i = 0; i < num_choices; i++) {
+	int result = 0;
+	for(int i = 0; i < num_choices; i++) {
 		char* cmp_str = va_arg(ap, char*);
 		if(strcmp(opt, cmp_str) == 0) {
 			result = i;
